Fix for Google auth on Chrome apps
After Chrome 47, launchWebAuthFlow no longer works with a
accounts.google.com URI, so we are unable to use it to auth with Google
( https://code.google.com/p/chromium/issues/detail?id=559523 )

This adds support code for a workaround where we redirect to a
redirect.html page hosted on the web, and from there use the
chrome.runtime APIs to communicate the auth response back to the
extension.

In order to use it developers will need to host the redirect.html file
with hello.js and set that as the redirect_url for hello.js, add the
chrome_ext_id option to their extension’s ID, and add an
‘externally_connectable’ option to their manifest to authorize their
site to send messages to their extension.