Add timeout cancelling to login
Hellojs opens up an iframe which redirects and communicates to the parent window using url hash
when there is a network error, redirect.html will never be loaded hence the promise can never
be rejected. Hello.js also has a recursive background timeout loop which checks for the status
of such window and from that it will resolve a promise callback stored in the window object and
remove th iframe from DOM. This loop will never detect the status at all during network error
hence the iframe is not removed from DOM.

Solution here is to add a failsafe that rejects promise and cleans up the iframe if the loop never
resolved the promise after a certain amount of time specified by the upstream login call.