Allow for positioning the auth popup while also giving dimensions
It seems that passing left and top in the options does nothing if you also pass width and height. When width and height is set, it will set left and top to position the window at the center of the screen, and ignore what's in the options. Add a condition that makes it only do that when left and top are not set in the options.

ps: using double equality '==' is not allowed